refactor(pipeline): move image config schema to pipeline/schemas/
Co-locates ImageConfig with DatasetConfig under pipeline/schemas/,
aligning all Pydantic config schemas in one place.

Refs #267
