+++ pipeline/schemas/image_config.py
